requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
aiohttp=3.8.3=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.2.0=pypi_0
blas=1.0=mkl
blobfile=2.0.1=pypi_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
contourpy=1.0.7=pypi_0
cryptography=38.0.4=py310h9ce1e76_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=12.0.76=0
cuda-documentation=12.0.76=0
cuda-driver-dev=11.7.99=0
cuda-gdb=12.0.90=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.0.78=0
cuda-nsight-compute=12.0.0=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=12.0.76=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=12.0.90=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=12.0.90=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=12.0.90=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
cycler=0.11.0=pypi_0
dall-e=0.1=pypi_0
datasets=2.9.0=pypi_0
dill=0.3.6=pypi_0
evaluate=0.4.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
fsspec=2023.1.0=pypi_0
gds-tools=1.5.0.59=0
giflib=5.2.1=h5eee18b_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
huggingface-hub=0.12.0=pypi_0
idna=3.4=py310h06a4308_0
iniconfig=2.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufft-dev=10.7.2.124=h98a8f43_0
libcufile=1.5.0.59=0
libcufile-dev=1.5.0.59=0
libcurand=10.3.1.50=0
libcurand-dev=10.3.1.50=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.8=h7f8727e_5
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=hecacb30_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
matplotlib=3.6.3=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
multidict=6.0.4=pypi_0
multiprocess=0.70.14=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
nsight-compute=2022.4.0.15=0
numpy=1.23.5=py310hd5efca6_0
numpy-base=1.23.5=py310h8e6c178_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
packaging=23.0=pypi_0
pandas=1.5.3=pypi_0
pillow=9.3.0=py310hace64e9_1
pip=22.3.1=py310h06a4308_0
pyarrow=11.0.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.9=h7a1cb2a_0
python-dateutil=2.8.2=pypi_0
pytorch=1.13.1=py3.10_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h67b0de4_1
pytorch-mutex=1.0=cuda
pytz=2022.7.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
regex=2022.10.31=pypi_0
requests=2.28.1=py310h06a4308_0
responses=0.18.0=pypi_0
scikit-learn=1.2.1=pypi_0
scipy=1.10.0=pypi_0
seaborn=0.12.2=pypi_0
setuptools=65.6.3=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.1=h5082296_0
tdqm=0.0.1=pypi_0
threadpoolctl=3.1.0=pypi_0
timm=0.6.12=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.2=pypi_0
torchaudio=0.13.1=py310_cu117
torchvision=0.14.1=py310_cu117
tqdm=4.64.1=pypi_0
transformers=4.26.0=pypi_0
typing_extensions=4.4.0=py310h06a4308_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.14=py310h06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
xxhash=3.2.0=pypi_0
xz=5.2.10=h5eee18b_1
yarl=1.8.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
